requirements.txt
transformers==4.30.2
pillow==10.0.0
gradio==3.40.1
torch==2.0.1
requests==2.31.0 